Updated cron job to use csv logs for updating or restting habit streaks at 12:00

+++ Backend/services/Habit_Services/batch_ops.py
@@ -1,13 +1,13 @@
 import json
-from pathlib import Path
 from typing import Dict, List
-from datetime import datetime
+from datetime import date, datetime, timedelta
 
-import variables.dirs as dirs
+from variables.paths import habits_dir, logs_dir
 from fastapi import HTTPException
 
 import logging
 from threading import Lock
+from services.Habit_Services.utils import parse_logs, update_habits
 
 logger = logging.getLogger("batch_ops")
 logging.basicConfig(
@@ -18,19 +18,18 @@ logging.basicConfig(
 class BatchOps:
     def __init__(self):
         self._all_habits: Dict = {}
-        self._dir: Path = dirs.habits_dir
         self._lock = Lock()
         self.load_all()
 
     def load_all(self):
         self._all_habits.clear()
 
-        for file in self._dir.rglob("*.json"):
+        for file in habits_dir.rglob("*.json"):
             if not file.is_file():
                 logger.debug(f"Skipping an object ({file.name}) thats not a file .")
                 continue
 
-            habits = List[dict]
+            habits: List[dict] = []
 
             if file.stat().st_size == 0:
                 logger.debug(f"Empty file {file.name} found .")
@@ -44,35 +43,43 @@ class BatchOps:
 
                             if not isinstance(habits, list):
                                 logger.error(
-                                    f"Habits from file {file.name} were not inside a valid list ."
+                                    f"Habit Object from file '{file.name}' was not a valid list ."
                                 )
-                                habits = []
+                                continue
 
                     except json.JSONDecodeError:
                         logger.error(
                             f"{file.name} contains invalid json so could not parse it ."
                         )
+                        continue
 
-            self._all_habits[file.stem] = habits
-
-    def reset_fields(self, h):
-        done = h["completed"]
-        fields = {
-            "completed": False,
-            "status": "incomplete",
-            "streak": h["streak"] + 1 if done else 0,
-        }
-        upd_habit = {**h, **fields}
-        return upd_habit
+            self._all_habits[file.name] = habits
 
     def update_streak(self):
         for f_name, habits in self._all_habits.items():
-            habits = [self.reset_fields(h) for h in habits if isinstance(h, dict)]
-            file = self._dir / f_name
+            habits_file = habits_dir / f_name
+
+            if not habits_file.exists():
+                logger.debug(
+                    f"Trying to read file with name: {f_name} that doesnt exist."
+                )
+                continue
+
+            logs_file = habits_file.name.split("_")[0] + "_logs.csv"
+            logs_file = logs_dir / logs_file
+
+            if not logs_file.exists():
+                # user has habits but no logs prob due to being offline mostly. Generate notif for the user.
+                continue
+
+            yesterday = date.today() - timedelta(days=1)
+            habit_logs: dict = parse_logs(logs_file, yesterday)
+
+            habits = update_habits(habits, habit_logs, yesterday, logs_file)
 
             with self._lock:
                 try:
-                    with open(f"{file}.json", "w") as f:
+                    with open(habits_file, "w") as f:
                         json.dump(habits, f)
                         logger.info(
                             f"Streaks reset successfully on {datetime.now().isoformat()}"
@@ -86,22 +93,10 @@ class BatchOps:
 
     def correct_habits(self):
         for f_name, habits in self._all_habits.items():
-            if len(habits) == 0:
-                logger.info(f"Habits for file {f_name} were empty !")
-                continue
+            file = habits_dir / f_name
 
-            incorrect_habits = [
-                habit
-                for habit in habits
-                if habit["completed"] and habit["status"] == "incomplete"
-            ]
+            for h in habits:
+                h["completed"] = False
 
-            last_completed = [h["last_completed"] for h in incorrect_habits]
-            try:
-                for l_c in last_completed:
-                    completed_today = (
-                        datetime.fromisoformat(l_c) == datetime.now().date()
-                    )
-
-            except Exception as e:
-                print(e)
+            with open(file, "w") as f:
+                json.dump(habits, f)
